requirements.txt
anthropic>=0.18.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
undetected-chromedriver>=3.5.0
selenium>=4.15.0
